renaming package nodebased -> npm
