+++ lib-extra/src/main/java/com/diffplug/spotless/extra/npm/prettier/PrettierFormatterStep.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.spotless.extra.nodebased.prettier;
+package com.diffplug.spotless.extra.npm.prettier;
 
 import static java.util.Arrays.asList;
 import static java.util.Objects.requireNonNull;
@@ -28,10 +28,10 @@ import com.diffplug.spotless.FormatterFunc;
 import com.diffplug.spotless.FormatterStep;
 import com.diffplug.spotless.Provisioner;
 import com.diffplug.spotless.ThrowingEx;
-import com.diffplug.spotless.extra.nodebased.NpmConfig;
-import com.diffplug.spotless.extra.nodebased.NpmFormatterStepStateBase;
-import com.diffplug.spotless.extra.nodebased.prettier.options.PrettierOptions;
-import com.diffplug.spotless.extra.nodebased.wrapper.*;
+import com.diffplug.spotless.extra.npm.NpmConfig;
+import com.diffplug.spotless.extra.npm.NpmFormatterStepStateBase;
+import com.diffplug.spotless.extra.npm.prettier.options.PrettierOptions;
+import com.diffplug.spotless.extra.npm.wrapper.*;
 
 public class PrettierFormatterStep {
