+++ lib-extra/src/main/java/com/diffplug/spotless/extra/npm/tsfmt/TsFmtOptions.java
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.spotless.extra.nodebased.tsfmt;
+package com.diffplug.spotless.extra.npm.tsfmt;
 
 import java.io.Serializable;
 import java.util.Optional;
 
-import com.diffplug.spotless.extra.nodebased.wrapper.NodeJSWrapper;
-import com.diffplug.spotless.extra.nodebased.wrapper.V8ObjectWrapper;
+import com.diffplug.spotless.extra.npm.wrapper.NodeJSWrapper;
+import com.diffplug.spotless.extra.npm.wrapper.V8ObjectWrapper;
 
 public class TsFmtOptions implements Serializable {
