+++ lib-extra/src/main/java/com/diffplug/spotless/extra/npm/tsfmt/TsFmtResult.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.spotless.extra.nodebased.tsfmt;
+package com.diffplug.spotless.extra.npm.tsfmt;
 
 public class TsFmtResult {
