+++ lib-extra/src/main/java/com/diffplug/spotless/extra/npm/wrapper/ReflectiveObjectWrapper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.spotless.extra.nodebased.wrapper;
+package com.diffplug.spotless.extra.npm.wrapper;
 
 import static java.util.Objects.requireNonNull;
