+++ lib-extra/src/main/resources/com/diffplug/spotless/extra/npm/prettier/package.json
