+++ lib-extra/src/main/resources/com/diffplug/spotless/extra/npm/tsfmt/package.json
