--- lib-extra/src/test/java/com/diffplug/spotless/extra/nodebased/NpmTest.java
@@ -1,5 +1,0 @@
-package com.diffplug.spotless.extra.nodebased;
-
-public interface NpmTest {
-
-}
