+++ lib-extra/src/test/java/com/diffplug/spotless/extra/npm/NpmTest.java
@@ -1,0 +1,5 @@
+package com.diffplug.spotless.extra.npm;
+
+public interface NpmTest {
+
+}
